+++ android/src/main/kotlin/dev/ltag/stone_payments/usecases/PaymentUsecase.kt
@@ -3,12 +3,16 @@ package dev.ltag.stone_payments.usecases
 import android.os.Handler
 import android.os.Looper
 import android.util.Log
+import com.google.gson.Gson
 import br.com.stone.posandroid.providers.PosPrintReceiptProvider
 import br.com.stone.posandroid.providers.PosTransactionProvider
 import dev.ltag.stone_payments.Result
 import dev.ltag.stone_payments.StonePaymentsPlugin
+import dev.ltag.stone_payments.core.Helper
+import dev.ltag.stone_payments.core.SerializableTransactionObject
 import io.flutter.plugin.common.MethodChannel
 import stone.application.enums.Action
+import stone.application.enums.EntryMode
 import stone.application.enums.InstalmentTransactionEnum
 import stone.application.enums.ReceiptType
 import stone.application.enums.TransactionStatusEnum
@@ -21,14 +25,14 @@ import stone.utils.Stone
 class PaymentUsecase(
     private val stonePayments: StonePaymentsPlugin,
 ) {
-    private val context = stonePayments.context;
+    private val context = stonePayments.context
 
     fun doPayment(
         value: Double,
         type: Int,
         installment: Int,
         print: Boolean?,
-        callback: (Result<String>) -> Unit,
+        callback: (Result<Boolean>) -> Unit,
     ) {
         try {
             stonePayments.transactionObject = TransactionObject()
@@ -36,12 +40,12 @@ class PaymentUsecase(
             val transactionObject = stonePayments.transactionObject
 
             transactionObject.instalmentTransaction =
-                InstalmentTransactionEnum.getAt(installment - 1);
-            transactionObject.typeOfTransaction =
-                if (type == 1) TypeOfTransactionEnum.CREDIT else TypeOfTransactionEnum.DEBIT;
-            transactionObject.isCapture = true;
-            val newValue: Int = (value * 100).toInt();
-            transactionObject.amount = newValue.toString();
+                InstalmentTransactionEnum.getAt(installment - 1)
+            transactionObject.typeOfTransaction = TypeOfTransactionEnum.values()[type]
+
+            transactionObject.isCapture = true
+            val newValue: Int = (value * 100).toInt()
+            transactionObject.amount = newValue.toString()
 
             val provider = PosTransactionProvider(
                 context,
@@ -63,7 +67,7 @@ class PaymentUsecase(
                                     PosPrintReceiptProvider(
                                         context, transactionObject,
                                         ReceiptType.MERCHANT,
-                                    );
+                                    )
 
                                 posPrintReceiptProvider.connectionCallback = object :
                                     StoneCallbackInterface {
@@ -74,7 +78,7 @@ class PaymentUsecase(
                                     }
 
                                     override fun onError() {
-                                        Log.d("ERRORPRINT", transactionObject.toString())
+                                        Log.d("ERROR", transactionObject.toString())
 
                                     }
                                 }
@@ -84,20 +88,17 @@ class PaymentUsecase(
                             }
                             sendAMessage("APPROVED")
 
-//                            val serializableTransaction = convertToSerializable(transactionObject);
-//
-//                            val jsonString = Json.encodeToString(serializableTransaction)
-                            callback(Result.Success(""))
+                            callback(Result.Success(true))
                         }
                         TransactionStatusEnum.DECLINED -> {
                             val message = provider.messageFromAuthorize
                             sendAMessage(message ?: "DECLINED")
-                            callback(Result.Success(""))
+                            callback(Result.Success(false))
                         }
                         TransactionStatusEnum.REJECTED -> {
                             val message = provider.messageFromAuthorize
                             sendAMessage(message ?: "REJECTED")
-                            callback(Result.Success(""))
+                            callback(Result.Success(false))
                         }
                         else -> {
                             val message = provider.messageFromAuthorize
@@ -113,7 +114,7 @@ class PaymentUsecase(
 
                     sendAMessage(provider.transactionStatus?.name ?: "ERROR")
 
-                    callback(Result.Error(Exception("ERROR")));
+                    callback(Result.Error(Exception("ERROR")))
                 }
 
                 override fun onStatusChanged(p0: Action?) {
@@ -126,7 +127,111 @@ class PaymentUsecase(
 
         } catch (e: Exception) {
             Log.d("ERROR", e.toString())
-            callback(Result.Error(e));
+            callback(Result.Error(e))
+        }
+
+    }
+
+    fun doTransaction(
+        value: Double,
+        type: Int,
+        installment: Int,
+        print: Boolean?,
+        callback: (Result<String>) -> Unit,
+    ) {
+        try {
+            stonePayments.transactionObject = TransactionObject()
+
+            val transactionObject = stonePayments.transactionObject
+
+            transactionObject.instalmentTransaction =
+                InstalmentTransactionEnum.getAt(installment - 1)
+            transactionObject.typeOfTransaction = TypeOfTransactionEnum.values()[type]
+
+            transactionObject.isCapture = true
+            val newValue: Int = (value * 100).toInt()
+            transactionObject.amount = newValue.toString()
+
+            val provider = PosTransactionProvider(
+                context,
+                transactionObject,
+                Stone.getUserModel(0),
+            )
+
+            provider.setConnectionCallback(object : StoneActionCallback {
+
+                override fun onSuccess() {
+                    when (val status = provider.transactionStatus) {
+                        TransactionStatusEnum.APPROVED -> {
+                            if (print == true) {
+                                val posPrintReceiptProvider =
+                                    PosPrintReceiptProvider(
+                                        context, transactionObject,
+                                        ReceiptType.MERCHANT,
+                                    )
+
+                                posPrintReceiptProvider.connectionCallback = object :
+                                    StoneCallbackInterface {
+
+                                    override fun onSuccess() {
+
+                                        Log.d("SUCCESS", transactionObject.toString())
+                                    }
+
+                                    override fun onError() {
+                                        Log.d("ERROR", transactionObject.toString())
+
+                                    }
+                                }
+
+                                posPrintReceiptProvider.execute()
+
+                            }
+                            sendAMessage("APPROVED")
+                        }
+                        TransactionStatusEnum.DECLINED -> {
+                            val message = provider.messageFromAuthorize
+                            sendAMessage(message ?: "DECLINED")
+                        }
+                        TransactionStatusEnum.REJECTED -> {
+                            val message = provider.messageFromAuthorize
+                            sendAMessage(message ?: "REJECTED")
+                        }
+                        else -> {
+                            val message = provider.messageFromAuthorize
+                            sendAMessage(message ?: status.name)
+                        }
+                    }
+
+                    val serializableTransaction = SerializableTransactionObject.from(transactionObject)
+
+                    val jsonString = Gson().toJson(serializableTransaction)
+                    callback(Result.Success(jsonString))
+                }
+
+                override fun onError() {
+
+                    Log.d("RESULT", "ERROR")
+
+                    sendAMessage(provider.transactionStatus?.name ?: "ERROR")
+
+                    callback(Result.Error(Exception("ERROR")))
+                }
+
+                override fun onStatusChanged(action: Action?) {
+                    if (action == Action.TRANSACTION_WAITING_QRCODE_SCAN) {
+                        sendQrCode(Helper().convertBitmapToString(transactionObject.qrCode))
+                    }
+                    sendAMessage(action?.name!!)
+                }
+            })
+
+            provider.execute()
+
+
+        } catch (e: Exception) {
+            Log.d("ERROR", e.toString())
+            callback(Result.Error(e))
         }
 
     }
@@ -138,6 +243,16 @@ class PaymentUsecase(
                 "stone_payments",
             )
             channel.invokeMethod("message", message)
+        }
+    }
+
+    private fun sendQrCode(qrCode: String) {
+        Handler(Looper.getMainLooper()).post {
+            val channel = MethodChannel(
+                StonePaymentsPlugin.flutterBinaryMessenger!!,
+                "stone_payments",
+            )
+            channel.invokeMethod("pixQrCode", qrCode)
         }
     }
 }